% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corFunAR1.R
\name{blockinv}
\alias{blockinv}
\title{Inversion of 2x2 block matrix}
\usage{
blockinv(M, .solve = solve, Dinv = lapply(M[c(1, 4)], .solve))
}
\arguments{
\item{M}{list of matrices of \code{dim(M) == 2} with matching columns / rows}

\item{Dinv}{list of inverses of \code{M[[1,1]]} and \code{M[[2,2]]} in that order.}
}
\description{
Inversion of 2x2 block matrix
}
