% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corFunAR1.R
\name{blockinv_symm}
\alias{blockinv_symm}
\title{Inversion of symmetric 2x2 block matrix}
\usage{
blockinv_symm(bdiag, odiag, .solve = solve, bdiaginv = lapply(bdiag, .solve))
}
\arguments{
\item{bdiag}{list of two square matrices \code{dim(M) == 2} forming the block diagonal}

\item{odiag}{matrix with \code{nrow(odiag) == nrow(bdiag[[1]])} and
\code{ncol(odiag) == ncol(bdiag[[2]])} forming the off diagonal block}

\item{bdiaginv}{list of inverses of \code{bdiag} (to allow their pre-computation).}
}
\description{
Inversion of symmetric 2x2 block matrix
}
