% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_tp_smoothing.R
\name{cov_cross_setup}
\alias{cov_cross_setup}
\title{Set up smooth covariance fitting for two cross-covariances}
\usage{
cov_cross_setup(X1, X2 = X1, S1, S2 = S1)
}
\arguments{
\item{X1, X2}{lists of according marginal design matrices}

\item{S1, S2}{lists with penalty matrix (or matrices) of marginal design matrices}
}
\value{
symmetric covariance fitting function returning coefficient matrix of
estimated tensor product smooth and taking the arguments
\code{y1,y2}: lists of zero-mean responses (typically model residuals) and
\code{sp1,sp2}: numeric (vectors) specifying the smoothing parameter(s) for the penalty.
}
\description{
Set up smooth covariance fitting for two cross-covariances
}
