% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_tp_smoothing.R
\name{cov_symm_setup}
\alias{cov_symm_setup}
\title{Set up smooth covariance fitting}
\usage{
cov_symm_setup(X, S, ridge = 0)
}
\arguments{
\item{X}{list of according marginal design matrices}

\item{S}{list with penalty matrix (or matrices) of marginal design matrices}

\item{ridge}{passed on to \code{\link{get_demmlerreinsch_solver}}: a relative
ridge on the pair design product, \code{0} by default.}
}
\value{
symmetric covariance fitting function returning coefficient matrix of
estimated tensor product smooth and taking the arguments
\code{y}: a list of zero-mean responses (typically model residuals) and
\code{sp}: a numeric (vector) specifying the smoothing parameter(s) for the penalty.
}
\description{
Set up smooth covariance fitting
}
