% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam.R
\name{fam_EM}
\alias{fam_EM}
\title{Run EM algorithm for estimation of timeseries FAM}
\usage{
fam_EM(
  init,
  gam_prefit,
  cov_smooth,
  id,
  cov_sp = 0,
  truncate = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Run EM algorithm for estimation of timeseries FAM
}
