% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam.R
\name{fam_fit}
\alias{fam_fit}
\title{Fit functional regression model from within \code{fam}}
\usage{
fam_fit(
  gam_prefit,
  cov_smooth,
  id,
  cov_sp = 0,
  quadrature_dat = NULL,
  quadrature_weights = 1/length(quadrature_dat[[1]]),
  truncate = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fit functional regression model from within \code{fam}
}
