% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.construct.symm.smooth.spec.R
\name{force_non_negative}
\alias{force_non_negative}
\title{Force symmetric smoothers to be non-negative definite}
\usage{
force_non_negative(object, term = NULL)
}
\arguments{
\item{object}{\code{gam} object containing smoother of class \code{symm.smooth}.}

\item{term}{label of smoother where the coefficients are to be modified. For the
default \code{term=NULL} the first smoother of class \code{symm.smooth} in the model is chosen.

Sets negative eigenvalues of the tensor product coefficient matrix to zero.}
}
\value{
a modified version of \code{object} with the coefficients of \code{term}
modified produce a non-negative definite symmetric smooth.
}
\description{
Force symmetric smoothers to be non-negative definite
}
