% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corSmooth.R
\name{get_demmlerreinsch_solver}
\alias{get_demmlerreinsch_solver}
\title{Title}
\usage{
get_demmlerreinsch_solver(XtX, S)
}
\arguments{
\item{XtX}{inner product matrix of design matrix columns}

\item{S}{penalty matrix}
}
\description{
Title
}
