% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corSmooth.R
\name{get_sp}
\alias{get_sp}
\title{Extract gam smoothing paramter estimates from gamm object}
\usage{
get_sp(var.param)
}
\arguments{
\item{var.param}{}
}
\value{
vector of smoothing parameters for gam object.
}
\description{
Internal helper function for \code{corMatrix.corSmooth}
}
