% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corSmooth.R
\name{get_XxXtYxY_noDiag}
\alias{get_XxXtYxY_noDiag}
\title{Title}
\usage{
get_XxXtYxY_noDiag(X, Y)
}
\arguments{
\item{X}{marginal designmatrix in \code{kronecker{X,X}}.}

\item{Y}{marginal response in \code{kronecker{Y,Y}} minus diagonal.}
}
\description{
Title
}
