% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globeplotter.R
\name{globeplotter_goode}
\alias{globeplotter_goode}
\title{Plot dataset with prediction in Goode holomosine projections}
\usage{
globeplotter_goode(
  pred = NULL,
  pred_variables = "TEMP",
  dat = NULL,
  dat_variables = NULL,
  dims = c("LONGITUDE", "LATITUDE"),
  return_fun = FALSE,
  method = c("Goode_ocean", "Goode_land"),
  pred_raster_args = list(palette = "-YlGnBu", n = 9, legend.reverse = TRUE, title =
    "Water temperature"),
  dat_dots_args = list(col = "darkred"),
  bgpolygon_borders_args = list(col = "darkgrey"),
  land_raster_args = list("elevation", palette = "-Greys", n = 9, legend.show = FALSE),
  world_lines_args = list(),
  graticule_lines_args = list(col = "darkgrey")
)
}
\arguments{
\item{dims}{character vector, names of longitude and latitude columns.}

\item{method}{plot method. Defaults to 'Goode_ocean', plotting the data
with R package \code{tmap} in the ocean-focussed version of the
the intersected holomosine projection of Goode (1925).}

\item{x}{data.frame containing the \code{variable} to plot and
respective \code{longitude}s and \code{latitude}s as columns.}

\item{variable}{character, name of variable to illustrate.}
}
\description{
Plot dataset with prediction in Goode holomosine projections
}
