% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globeplotter.R
\name{goode}
\alias{goode}
\title{prepare utilities for plotting Goode holomosine projections (with tmap)}
\usage{
goode(type = c("ocean", "land"), longrat = 30, latgrat = longrat/2)
}
\arguments{
\item{type}{the ocean (default) or land version of the projection}
}
\value{
list containing different utlities
}
\description{
prepare utilities for plotting Goode holomosine projections (with tmap)
}
