% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDynamic.R
\name{Initialize.corDynamic_init}
\alias{Initialize.corDynamic_init}
\title{Initialization of a dynamic correlation structure}
\usage{
\method{Initialize}{corDynamic_init}(object, data, ...)
}
\description{
Correlations structures of class \code{corDynamic} are not intended for
direct use but provide infrastructure for subclasses which require access to
objects which are not usually passed to it.
In particular, the initialization method provides the environment of the
\code{lme.formula} call executing it and to model residuals computed from
the last call.
}
