% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cov_data.R
\name{make_cov_data}
\alias{make_cov_data}
\title{Generate dataset for covariance estimation of irregular functional data}
\usage{
make_cov_data(data, response_var, id_var = NULL, sep = "")
}
\arguments{
\item{data}{a data.frame}

\item{response_var}{character string indicating the name of the response in \code{data}.}

\item{id_var}{character string indicating the name of the id variable in \code{data}
identifying the functional observations.}

\item{sep}{character string seperating argument/covariate names and their number 1 and 2.}
}
\value{
data.frame with data for covariance estimation
}
\description{
Generate dataset for covariance estimation of irregular functional data
}
