% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDynamic.R
\name{corDynamic}
\alias{corDynamic}
\title{Dynamic correlation structure}
\usage{
corDynamic(value = NULL, form = formula(value), fixed = FALSE)
}
\arguments{
\item{value}{a \code{corStruct} object indicating the working correlation
structure assumed initially when computing the dynamic correlation structure
subclass requires pre-fitting the model. The default \code{NULL} indicates working independence.}

\item{fixed}{an optional logical value indicating whether the coefficients
should be allowed to vary in the optimization, or kept fixed at their initial value.
Defaults to \code{FALSE}, in which case the coefficients are allowed to vary.}

\item{formula}{a formula}
}
\description{
Dynamic correlation structure
}
