% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corFunAR1.R
\name{corFunAR1}
\alias{corFunAR1}
\title{Functional AR-1 covariance structure based on covariance smoothing}
\usage{
corFunAR1(
  value = 0,
  form = ~1,
  fixed = FALSE,
  working_correlation = corCAR1,
  working_control = list(),
  verbose = FALSE
)
}
\arguments{
\item{value}{smoothing parameter for symmetric tensor-product smoother (\code{gam} smoother class \code{symm.smooth})}

\item{form}{a one sided formula of the form ~ t, or ~ t | g,
specifying a time covariate t and, optionally, a grouping factor g.
When a grouping factor is present in form, the correlation structure is assumed
to apply only to observations within the same grouping level;
observations with different grouping levels are assumed to be uncorrelated.}

\item{fixed}{an optional logical value indicating whether the coefficients
should be allowed to vary in the optimization, or kept fixed at their initial value.
Defaults to \code{FALSE}, in which case the coefficients are allowed to vary.}
}
\value{
an object of class \code{corFunAR1}, representing an covariance smoother autocorrelation structure.
}
\description{
Functional AR-1 covariance structure based on covariance smoothing
}
