% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam.R
\name{fam_init}
\alias{fam_init}
\title{Fit functional regression model from within \code{fam}}
\usage{
fam_init(
  gam_prefit,
  cov_smooth,
  id,
  cov_sp = 0,
  quadrature_dat = NULL,
  quadrature_weights = 1/length(quadrature_dat[[1]]),
  truncate = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
This returns OLS mean estimates underlying also
all estimates of covariance components.
}
