% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globeplotter.R
\name{globeplotter_3D}
\alias{globeplotter_3D}
\title{Plot data.frame with predictions in interactive 3D chart}
\usage{
globeplotter_3D(
  pred = NULL,
  pred_variables = "TEMP",
  dat = NULL,
  dat_variables = NULL,
  dims = c("LONGITUDE", "LATITUDE"),
  return_fun = FALSE,
  pred_args = list(palette = "-YlGnBu", n = 9, legend.reverse = TRUE, title =
    "Water temperature"),
  dat_args = list(col = "darkred"),
  land_args = list("elevation", palette = "-Greys", n = 9, legend.show = FALSE),
  world_args = list(),
  graticule_args = list(col = "darkgrey"),
  ...
)
}
\description{
Plot data.frame with predictions in interactive 3D chart
}
