% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_tp_smoothing.R
\name{predict_square_smooth}
\alias{predict_square_smooth}
\title{Predict a single-tensor-product smooth}
\usage{
predict_square_smooth(theta, X, decompose = FALSE, Gramian = NULL, ...)
}
\arguments{
\item{theta}{an object representing the estimated coefficients}

\item{decompose}{logical, should prediction be returned in form of its SVD?}

\item{Gramian}{an optional Gramian providing the Gam-matrix of the splines for SVD.
For NULL, the default, the SVD is conducted with respect to an inner product in
which the basis functions are orthonormal.}

\item{smooth_obj}{a gam smoother object with a \code{Predict.matrix} method}

\item{newdata}{data for prediction}
}
\description{
Predict a single-tensor-product smooth
}
