% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.construct.symm.smooth.spec.R
\name{smooth.construct.symm.smooth.spec}
\alias{smooth.construct.symm.smooth.spec}
\title{Symmetric bivariate smooths constructor}
\usage{
\method{smooth.construct}{symm.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{is a smooth specification object or a smooth object.}

\item{data}{a data frame, model frame or list containing the values
of the (named) covariates at which the smooth term is to be evaluated.}

\item{knots}{an optional data frame supplying any knot locations
to be supplied for basis construction.}
}
\value{
An object of class "symm.smooth". See \code{\link[mgcv]{smooth.construct}} for the elements it will contain.
}
\description{
The \code{symm} class is a smooth class that is appropriate for symmetric smooths, e.g. of covariance functions,
using tensor-product smooths in a \code{gam} formula. A constraint matrix is constructed
(see \code{\link[sparseFLMM]{make_summation_matrix}}) to impose
a (skew-)symmetry constraint on the smooth's coefficients,
which considerably reduces the number of coefficients that have to be estimated.
}
\details{
By default a symmetric bivariate smooth \eqn{g} is specified,
in the sense that \eqn{g(s, t) = g(t, s)}.
In contrast to the original implementation of the function in the package
\code{sparseFLMM}, this implementation also works for more general smooths
and any even number of arguments, i.e. \eqn{g(s1, s2, ..., t1, t2, ...) = g(t1, t2, ..., s1, s2, ...)}.

Several options can be specified via the argument \code{xt} of \code{s()} with the
default given by

\code{s(..., bs = "symm", xt = list(bsmargin = "ps", skew = FALSE, absorb.cons = TRUE, kroneckersum = TRUE))}

with
\itemize{
\item \code{bsmargin}: the marginal basis type (the \code{bs} argument used to construct them).
\item \code{skew}: should a skew-symmetric (or anti-symmetric) smooth with \eqn{g(s, t) = -g(t, s)}
be specified instead of a symmetric smooth?
\item \code{absorb.cons}: should the automatic sum-to-zero constraint be applied?
Turning it off, no additional transformation of the design and
penalty matrices is conducted and the obtained coefficients directly correspond
to them. In this case, the constant intercept should be excluded if \code{skew==FALSE}.
For \code{skew==TRUE}, the constraint is always off.
\item \code{kroneckersum}: should a Kronecker sum penalty be used instead of a Kronecker product penalty?
}

The underlying procedure is the following: First, the marginal spline design matrices and the corresponding
marginal difference penalties are built. Second, the tensor product of the marginal design matrices is built
and the bivariate penalty matrix is set up. Third, the constraint matrix is applied
to the tensor product design matrix and to the penalty matrix.
}
\examples{

library(mgcv)
library(funSphere)

# check out example dataset ------------------------------------------------

data(world.cities, package = "maps")

library(ggplot2)
library(patchwork)

caps <- world.cities[as.logical(world.cities$capital), ]

world <- map_data("world")

(p <- ggplot(data = NULL) +
    geom_map(
      data = world, map = world,
      aes(map_id = region),
      color = "black", fill = "lightgray", size = 0.1
    ) +
    coord_cartesian(xlim = c(-180, 180), ylim = c(-90, 90), expand = F) +
    coord_fixed()
) +
  geom_point(data = caps, aes(long, lat, col = sqrt(pop)))


# first fit strange symmetric 2D GP -------------------------------------
m0 <- gam(sqrt(pop) ~ s(long, lat, bs = "symm",
                        xt = list(bsmargin = "gp2"), m = -3), data = caps)
predgrid <- expand.grid(long = seq(-180, 180, 10), lat = seq(-180, 180, 10))
predgrid$pop0 <- predict(m0, predgrid)^2

p + geom_raster(data = predgrid, aes(long, lat, fill = sqrt(pop0)), alpha = .96) +
  scale_fill_viridis_c()

pop0 <- matrix(predgrid$pop0, nrow = sqrt(nrow(predgrid)))
all.equal(pop0, t(pop0))


# now try covariance data -------------------------------------------------
cits <- world.cities[, -2]
set.seed(8934)
n_i <- 30
n_sample <- 50
cits$sample <- NA
ids <- seq_len(nrow(cits))
for(i in n_sample)
  cits$sample[sample(ids, size = n_i, replace = FALSE)] <- i
cits <- cits[!is.na(cits$sample), ]
cits$pop <- as.double(cits$pop)

covcaps <- make_cov_data(cits, response_var = "pop")
nrow(covcaps)

m1 <- gam(sqrt(poppop) ~ s(long1, lat1, long2, lat2, bs = "symm",
                        xt = list(bsmargin = "gp2"), m = -3), data = covcaps)

predgrid1 <- expand.grid(
  long1 = seq(-180, 180, 10), lat1 = seq(-90, 90, 5),
  long2 = c(-90, 0), lat2 = c(-20,0) )
predgrid2 <- predgrid1
names(predgrid2) <- names(predgrid1)[c(3:4,1:2)]

predgrid1$poppop <- predict(m1, predgrid1)^2
predgrid2$poppop <- predict(m1, predgrid2)^2

(p + geom_raster(data = predgrid1,
                aes(long1, lat1,
                    fill = sqrt(poppop)),
                alpha = .96) +
  scale_fill_viridis_c() +
  facet_grid(lat2~long2, labeller = label_both)) +
(p + geom_raster(data = predgrid2,
                aes(long2, lat2,
                    fill = sqrt(poppop)),
                alpha = .96) +
  scale_fill_viridis_c() +
  facet_grid(lat1~long1, labeller = label_both))

predgrid_sym <- expand.grid(
  long1 = seq(-180, 180, 10), lat1 = c(-20,0),
  long2 = seq(-180, 180, 10), lat2 = c(-20,0) )
predgrid_sym$poppop <- predict(m1, predgrid_sym)^2

ggplot(predgrid_sym, aes(long1, long2, fill = sqrt(poppop))) +
  geom_raster() + geom_contour(aes(z = sqrt(poppop)), col = "white") +
  scale_fill_viridis_c() +
  facet_grid(lat1 ~ lat2, labeller = label_both) +
  coord_fixed()

predgrid_diag <- expand.grid(long1 = seq(-180, 180, 10),
                             lat1 = seq(-90, 90, 10))
predgrid_diag[c("long2", "lat2")] <- predgrid_diag[, 1:2]

predgrid_diag$poppop <- predict(m1, predgrid_diag)^2

p + geom_raster(data = predgrid_diag,
                aes(long1, lat1,
                    fill = sqrt(poppop)),
                alpha = .9) +
  scale_fill_viridis_c()


}
\references{
Cederbaum, Scheipl, Greven (2016): Fast symmetric additive covariance smoothing.
Stoecker, Pfeuffer, Steyer, Greven (2022): Elastic Full Procrustes Analysis via Hermitian Covariance Smoothing.
}
\seealso{
\code{\link[mgcv]{smooth.construct}} and \code{\link[mgcv]{smoothCon}} for details on constructors
}
\author{
Almond Stoecker, Jona Cederbaum
}
